requirements.txt
Adafruit-Blinka==6.12.0
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-mcp3xxx==1.4.5
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
aiohttp==3.5.4
aioresponses==0.7.2
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
crcmod==1.7
cycler==0.10.0
dsmr-parser==0.21
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
meld3==1.0.2
multidict==4.5.2
numpy==1.17.1
pandas==1.0.5
psycopg2==2.8.3
pythoncrc==1.21
PyFronius==0.4.6
pyftdi==0.53.2
pymongo==3.4.0
pyparsing==2.4.2
pyserial==3.4
pyserial-asyncio==0.4
pysma==0.6.5
python-dateutil==2.8.0
pytz==2019.2
pyusb==1.1.0
requests==2.22.0
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.7
supervisor==4.0.4
sysv-ipc==1.1.0
urllib3==1.25.3
Werkzeug==0.15.5
yarl==1.3.0
pytimeparse=1.1.8
